pyproject.toml
[project]
name = "podcastvox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.6",
    "anyio>=4.9.0",
    "fastapi>=0.115.12",
    "gradio>=5.32.0",
    "litellm>=1.72.0",
    "markitdown[pdf]>=0.1.2",
    "pydantic>=2.11.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.12",
    "ty>=0.0.1a7",
]
